Removed generation of intermediate object files
Object files, while useful in large projects so that chunks of code don't have
to be recompiled, are largely redundant in a project as small as the current
one. They also had to be retained, using `.SECONDARY:` on the `Makefile`, so
that symbolic information about executables could be accessed. However,
retaining object files introduced a subtle bug - because different matrix
implementations were used to generate `matrix.o`, if `matrix.o` already existed
(which it was with `.SECONDARY:`), the new matrix implementations wouldn't be
used. This meant that only a single matrix implementation was being used for
timings at any given time.

This change removed the `.SECONDARY:` line from the `Makefile` and instead
generates the executable directly from sources without using object files,
meaning that object files are not needed but yet symbolic information is
retained.